Fix the type of height

+++ 9-binary_tree_height.c
@@ -1,7 +1,6 @@
 #include "binary_trees.h"
 #include <stdlib.h>
 #include <stdio.h>
-#include <stddef.h>
 /**
  * binary_tree_height - hieght of a tree
  * @tree: pointer to the root node of the tree to measure the height.
@@ -9,7 +8,7 @@
  */
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int height = 0;
+	size_t height = 0;
 
 	if (tree == NULL)
 		return (0);
